requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
mitmproxy==10.1.5
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
tabulate==0.9.0
openapi-schema-pydantic==1.2.4
pyyaml==6.0.1
colorama==0.4.6
spacy==3.7.2
jinja2==3.1.2